txt files only

+++ klofr.py
@@ -174,16 +174,19 @@ async def autocorrect(ctx, query:str="None", number:str="1", *, separator:str=" 
 async def autocorrect_file(ctx, file: discord.Attachment, number:str="1", *, separator:str=" "):
     await ctx.defer()
 
-    uploaded_file = f"text_files/{file.filename}"
-    await file.save(uploaded_file)
+    if file.filename.endswith(".txt"):
+        uploaded_file = f"text_files/{file.filename}"
+        await file.save(uploaded_file)
 
-    try:
-        with open(uploaded_file, "r", encoding="utf-8") as input_file:
-            query = input_file.read()
-            msg = await prettify_autocorrector(query, int(number), separator)
-        os.remove(uploaded_file)
-    except Exception as e:
-        msg = e
+        try:
+            with open(uploaded_file, "r", encoding="utf-8") as input_file:
+                query = input_file.read()
+                msg = await prettify_autocorrector(query, int(number), separator)
+            os.remove(uploaded_file)
+        except Exception as e:
+            msg = e
+    else:
+        msg = "please only upload txt files"
 
     await ctx.send(msg)
